add comments and change some virables
