+++ master/jobmaster/benchmark/build.go
@@ -1,6 +1,7 @@
 package benchmark
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 
@@ -9,8 +10,8 @@ import (
 	"github.com/hanfei1991/microcosom/pkg/autoid"
 )
 
-// BUild JobMaster for benchmark workload.
-func BuildBenchmarkJobMaster(rawConfig string, idAllocator *autoid.Allocator, resourceMgr cluster.ResourceMgr, client cluster.ExecutorClient) (*Master, error) {
+// BuildBenchmarkJobMaster for benchmark workload.
+func BuildBenchmarkJobMaster(ctx context.Context, rawConfig string, idAllocator *autoid.Allocator, resourceMgr cluster.ResourceMgr, client cluster.ExecutorClient) (*Master, error) {
 	config, err := configFromJson(rawConfig)
 	if err != nil {
 		return nil, err
